+++ task-sdk/src/airflow/sdk/bases/decorator.py
@@ -26,7 +26,6 @@ from functools import cached_property, update_wrapper
 from typing import TYPE_CHECKING, Any, ClassVar, Generic, ParamSpec, Protocol, TypeVar, cast, overload
 
 import attr
-import libcst as cst
 import typing_extensions
 
 from airflow.sdk import timezone
@@ -38,6 +37,7 @@ from airflow.sdk.bases.operator import (
     parse_retries,
 )
 from airflow.sdk.definitions._internal.contextmanager import DagContext, TaskGroupContext
+from airflow.sdk.definitions._internal.decorators import remove_task_decorator
 from airflow.sdk.definitions._internal.expandinput import (
     EXPAND_INPUT_EMPTY,
     DictOfListsExpandInput,
@@ -684,44 +684,3 @@ def task_decorator_factory(
         )
 
     return cast("TaskDecorator", decorator_factory)
-
-
-class _TaskDecoratorRemover(cst.CSTTransformer):
-    def __init__(self, task_decorator_name: str) -> None:
-        self.decorators_to_remove: set[str] = {
-            "setup",
-            "teardown",
-            "task.skip_if",
-            "task.run_if",
-            task_decorator_name.strip("@"),
-        }
-
-    def _is_task_decorator(self, decorator_node: cst.Decorator) -> bool:
-        decorator_expr = decorator_node.decorator
-        if isinstance(decorator_expr, cst.Name):
-            return decorator_expr.value in self.decorators_to_remove
-        if isinstance(decorator_expr, cst.Attribute) and isinstance(decorator_expr.value, cst.Name):
-            return f"{decorator_expr.value.value}.{decorator_expr.attr.value}" in self.decorators_to_remove
-        if isinstance(decorator_expr, cst.Call):
-            return self._is_task_decorator(cst.Decorator(decorator=decorator_expr.func))
-        return False
-
-    def leave_FunctionDef(
-        self, original_node: cst.FunctionDef, updated_node: cst.FunctionDef
-    ) -> cst.FunctionDef:
-        new_decorators = [dec for dec in updated_node.decorators if not self._is_task_decorator(dec)]
-        if len(new_decorators) == len(updated_node.decorators):
-            return updated_node
-        return updated_node.with_changes(decorators=new_decorators)
-
-
-def remove_task_decorator(python_source: str, task_decorator_name: str) -> str:
-    """
-    Remove @task or similar decorators as well as @setup and @teardown.
-
-    :param python_source: The python source code
-    :param task_decorator_name: the decorator name
-    """
-    source_tree = cst.parse_module(python_source)
-    modified_tree = source_tree.visit(_TaskDecoratorRemover(task_decorator_name))
-    return modified_tree.code
